pyproject.toml
[project]
name = "proovy_ai"
version = "0.1.0"
description = "FastAPI + LangGraph starter chatbot service"
readme = "README.md"
authors = [{ name = "Proovy" }]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "langgraph>=0.2.30",
    "langchain>=0.2.17",
    "langchain-openai>=0.1.23",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "uvicorn>=0.30",
    "python-dotenv>=1.2.1",
    "langfuse>=3.11.2",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"
